requirements.txt
# ===========================================================
# RuoYi 平台接口自动化测试框架
# Python 依赖包清单
# ===========================================================
# 作者：ruoyi-test Team
# 版本：v1.0.0
# ===========================================================
# 安装命令：pip install -r requirements.txt
# ===========================================================

# 测试框架核心
pytest ~= 8.3.0
pytest-rerunfailures ~= 14.0

# 接口请求
requests ~= 2.32.0

# 报告生成
allure-pytest ~= 2.13.0

# 数据驱动
PyYAML ~= 6.0.2

# 数据库连接（MySQL）
PyMySQL ~= 1.1.1
